requirements.txt
flask
flask_sqlalchemy
flask-migrate
flask-jwt-extended
flask-argon2
python-frontmatter